Build fixes after rebase
