+++ Data/Tools/ezEditor/Localization/en/ScenePlugin.txt
@@ -381,6 +381,7 @@ Fresnel;Fresnel
 Refraction;Refraction
 SetColorMode;Set Color Mode
 ezBakedProbesComponent;Baked Probes
+ezBakedProbesVolumeComponent;Baked Probes Volume
 ezRopeRenderComponent;Rope Renderer;;https://ezengine.net/pages/docs/effects/rope-render-component.html
 ezFakeRopeComponent;Fake Rope Simulation;;https://ezengine.net/pages/docs/effects/fake-rope-component.html
 ezClothSheetComponent;Cloth Sheet;;http://ezengine.net/pages/docs/effects/cloth-sheet-component.html
